+++ CHANGELOG.md
@@ -9,9 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### To add
 
-- get user's conversations
-- get selected user's conversations
-- find the one in common
+- find the one in common (conversationWithUserIds mutation)
 - if no match, create the convo
 - else, open page at this convo
 
@@ -23,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Search bar to search users by firstname, lastname, service name or poste name
 - corresponding queries
 - modal populated with results
+- Conversation resolver hooks
 
 ## [v0.0.3](https://github.com/benzinho75/JEECE-collab/compare/v0.0.2...v0.0.3) - 2021-03-04